// src/app/features/auth/interfaces/quiz.interface.ts
export interface QuizResponse {
    userId: string;
    experience: string;
    domain: string;
    languages: string[];
    learningMethod: string;
    goal: string;
    timePerWeek: string;
    resources: string[];
    paymentPreference: string;
    topics: string[];
    discovery: string;
    createdAt: Date;
} 